Only compile widget on Darwin + AMD64
