+++ lib/caseymrm/menuet/startup.go
@@ -1,5 +1,4 @@
 //go:build darwin
-// +build darwin
 
 package menuet
