+++ lib/jotaen/terminalformat/init_windows.go
@@ -1,5 +1,4 @@
 //go:build windows
-// +build windows
 
 package terminalformat
