+++ src/app/mac_widget/run_dummy.go
@@ -1,0 +1,9 @@
+//go:build !(darwin && amd64)
+
+package mac_widget
+
+func IsWidgetAvailable() bool {
+	return false
+}
+
+func Run(forceRunThroughLaunchAgent bool) {}
