+++ src/app/util_unix.go
@@ -1,5 +1,4 @@
 //go:build !windows
-// +build !windows
 
 package app
